### 2025-04-30 #8139 @editorialbot
[PRE REVIEW]: Terragon: A Unified Framework for Earth Observation Data Cube Generation
**Submitting author:** @drnhhl (<a href="http://orcid.org/0000-0003-3380-4489">Adrian Höhl</a>)
**Repository:** https://github.com/drnhhl/terragon
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.0
