### 2025-04-30 #8139 @kthyng
@drnhhl I see you have example notebooks. Have you thought about setting up something like readthedocs to compile those notebooks into more formal documentation? This would be a nice addition to your software.
