+++ paper/paper.bib
@@ -43,7 +43,7 @@ pages = {18-27},
 year = {2017},
 note = {Big Remotely Sensed Data: tools, applications and experiences},
 issn = {0034-4257},
-doi = {https://doi.org/10.1016/j.rse.2017.06.031},
+doi = {10.1016/j.rse.2017.06.031},
 url = {https://www.sciencedirect.com/science/article/pii/S0034425717302900},
 author = {Noel Gorelick and Matt Hancher and Mike Dixon and Simon Ilyushchenko and David Thau and Rebecca Moore},
 keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization, Earth Engine},
