### 2025-04-30 #8139 @drnhhl
Hi @kthyng, thank you for guiding the process. We have a documentation on readthedocs, you can find it [here](https://terragon.readthedocs.io/en/latest/) or by clicking on the badge in the ReadMe.
I have also fixed the doi which was flagged by the editorialbot.
