### 2025-04-30 #8139 @drnhhl
I have looked through the [list](https://reviewers.joss.theoj.org/reviewers/search?area_id=6&keywords=remote+sensing&language=26&name=) and have filtered for persons with a remote sensing background who might be interested in reviewing:
trchudley
dostuffthatmatters
micha-silver
yotarazona
tobijjah
