### 2025-06-23 #8139 @jatkinson1000
Hi @drnhhl thanks for this submission to JOSS.

Before I start researching for reviewers I had a couple of questions to help me understand the scope and objectives of this work. It would be really helpful if you could clarify these for me.

- You currently support 4 APIs - what is the rationale behind choosing these, and what else might you support in the future? Is there a roadmap for this project?
    I ask as you (fairly) critiscise other packages for being single-data-source, so clarity of support to rectify this in your code would be useful. I see support for all CDSE and ASF is not complete yet.
    Whilst I acknowledge that there will always be more data sources, JOSS asks that [submissions are feature-complete, not half-baked](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) and I wonder if there is further work to be done here?
- You provide citations for some other packages, including a passing mention of [cubo](https://github.com/ESDS-Leipzig/cubo). However, at a glance the API looks to be almost identical, and there is some overlap in supported datasets. Is there a reason that this work is not a contribution to expand cubo?
- JOSS asks for [past or ongoing research projects using the software and recent scholarly publications enabled by it](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) to be included. You state that "various research groups actively utilize" Terragon, but without further elaboration. Can you expand on this in the paper?

Thanks!
