### 2025-06-26 #8139 @drnhhl
Hi @jatkinson1000,

thank you for taking on the editing for our paper. We are happy to address your questions.

- We selected the four APIs based on their prominence and widespread use in the Earth observation community. As you mentioned, while there are many additional APIs, adding more APIs would still leave a lot of APIs unaddressed. Our goal was to support the most important APIs rather than to be exhaustive. Hence, we think it is as "complete" as possible, since adding more APIs does not enhance completeness.
The ASF support focuses on currently active and widely used datasets. The omitted datasets are legacy products with limited relevance and are typically only available in specific regions. Regarding CDSE, only the TIFF-based collections are supported right now, as others use the NetCDF format. CDSE is planning to add STAC downloading support for all collections. At the moment, it only supports Sentinel 1 and 2 (see [here](https://documentation.dataspace.copernicus.eu/APIs/newSTACcatalogue.html#available-collections)). We intend to extend support as CDSE finalizes STAC access to further collections.
We are actively engaging with the community, and user feedback will guide future development. Thus, while the tool may evolve, we believe the current version meets the JOSS requirement for a feature-complete, robust tool.
- Although there is some functional overlap, our design diverges substantially from Cubo’s. Cubo builds fixed-size data cubes around a centroid point using UTM zones, while Terragon supports arbitrary polygons as input and preserves the source CRS, avoiding pixel misalignment and overlaps. We aim to have a cube that closely mimics the shape of the polygon, making it a far more flexible approach.
Additionally, we utilize other software packages to access data providers. For example, while Cubo uses Xee for Google Earth Engine access, we use Geedim, which has several advantages for our approach. Adapting Cubo to support this paradigm would require rewriting its core architecture. We believe Terragon constitutes a distinct contribution rather than an extension of Cubo.
- While there are currently no published studies using Terragon, it is employed in at least two manuscripts that are under review. We have not cited them in the paper due to their unpublished status, but the datasets produced using Terragon are publicly available: [CropClimateX](https://huggingface.co/datasets/torchgeo/CropClimateX), [Sen12Landslides](https://huggingface.co/datasets/paulhoehn/Sen12Landslides). Furthermore, we are aware of several people preparing data with the software, which is one reason we want to submit the paper to JOSS.

We hope this clarifies our design choice and scope. Please let us know if you need any further details.
