+++ paper/paper.md
@@ -22,7 +22,7 @@ affiliations:
    index: 2
  - name: Munich Center for Machine Learning
    index: 3
-date: 28 April 2025
+date: 30 June 2025
 bibliography: paper.bib
 
 ---
@@ -32,14 +32,14 @@ Terragon (Earth(Poly)gon) is a Python package facilitating access to remote sens
 
 ![Terragon Workflow \label{workflow}](image.pdf){height="300pt"}
 
-Currently, several common data providers are supported, including Google Earth Engine [@GORELICK201718], Planetary Computer [@microsoft_open_source_2022_7261897], Copernicus Data Space Ecosystem [@ecosystemCopernicusData2024] and Alaska Satellite Facility [@ASFHome]. The goal is to further develop and maintain this tool by incorporating additional data providers and implementing processing techniques, such as mosaicking and resampling, based on community needs. Various research groups actively utilize the proposed solution for data preparation purposes.
+Currently, several common data providers are supported, including Google Earth Engine [@GORELICK201718], Planetary Computer [@microsoft_open_source_2022_7261897], Copernicus Data Space Ecosystem [@ecosystemCopernicusData2024] and Alaska Satellite Facility [@ASFHome]. The goal is to further develop and maintain this tool by incorporating additional data providers and implementing processing techniques, such as mosaicking and resampling, based on community needs. The software was leveraged to prepare two large-scale datasets Sen12Landslides [@Sen12Landslides_2025] and CropClimateX [@cropclimatex_dataset]. Sen12Landslides contains 75,000 landslide annotations and has over 12,000 patches from Sentinel 1 and 2. CropClimateX contains 15,500 small data cube spanning 1527 counties in the USA, it spans multiple sensors, weather and extreme events, soil and terrain features. More projects are in preparation.
 
 Overall, Terragon reduces the resources required for the time-consuming process of accessing and downloading data from various APIs, combining it into a consistent, reusable and cost-effienct framework. 
 
 # Statement of need
 The number of active earth-observation satellites is increasing [@SatelliteDatabase], and platforms like Sentinel-2 are generating vast amounts of data. This development has significantly impacted remote sensing analysis by providing access to large-scale datasets, both free and commercial. Simultaneously, there is an increasing demand for analysis-ready data to develop data-driven methods, such as machine learning models. However, the process of gathering this data remains time-consuming. Existing platforms are fragmented, often providing only specific satellites, formats, and costum APIs. As a result, this leads to repeated work and poor compatibility across platforms, highlighting the need for more convenient access and a unified, straightforward framework.
 
-Although many useful tools have already been developed to help people access and process Earth observation data [@Taconet2024; @Chudley2024; @Haan2023; @montero2024cubo; @hogenson2020hyp3], these tools often support only one satellite, product, and API, or are built for specific programming languages like R. This limits their broader applicability, especially for researchers looking to combine data from multiple sources or work within a Python-based scientific stack, where Python has become the standard for machine learning research.
[…]
