+++ paper/paper.bib
@@ -87,3 +87,20 @@ keywords = {Cloud computing, Big data, Analysis, Platform, Data democratization,
   doi = {10.5194/esd-11-201-2020},
   langid = {english}
 }
+
+@article{cropclimatex_dataset,
+  title = {{CropClimateX}: A Large-Scale, Multitask, Multisensory Dataset for Crop Monitoring under Climate Extremes},
+  author = {H{\"o}hl, Adrian and Ofori-Ampofo, Stella and Fern{\'a}ndez-Torres, Miguel-{\'A}ngel and Kuzu, Ridvan Salih and Zhu, Xiao Xiang},
+  year = {2025},
+  journal = {Hugging Face},
+  doi = {10.57967/hf/5047},
+}
+
+@misc{Sen12Landslides_2025,
+	author       = { H{\"o}hn, Paul and Heidler, Konrad and Behling, Robert and Zhu, Xiao Xiang },
+	title        = { Sen12Landslides - A spatio-temporal Dataset for satellite-based Landslide Detection },
+	year         = 2025,
+	url          = { https://huggingface.co/datasets/paulhoehn/Sen12Landslides },
+	doi          = { 10.57967/hf/5883 },
+	publisher    = { Hugging Face }
+}
