### 2025-07-07 #8139 @drnhhl
@usethedata

Thank you for your thoughtful and encouraging comment, as well as your insights into NASA DAACs. We have seen the API and agree it would be a valuable resource. We are planning to integrate it in the near future. We greatly appreciate your expertise in this area, and we will certainly keep it in mind as we move forward, if we need guidance or support with the integration.
