### 2025-07-22 994bcda
WIP: first steps earthaccess
