### 2025-08-01 88be4b8
add metadata for pc
