### 2025-08-04 f5547d7
Update earthaccess tests
