### 2025-08-04 c65d4fb
Implement/Update first functions
