### 2025-08-11 #8139 @drnhhl
Good morning @jatkinson1000
I am not fully familiar with the journal’s process and was wondering if a sufficient number of reviewers have been secured. If there is any information or action required from our side to help move things forward, we would be happy to assist.
