### 2025-08-11 #8139 @jatkinson1000
Hi @drnhhl I have been on leave so slower responses here of late.

JOSS requires 2 reviewers, with a strong preference for 3, so we still need to find another two.
As you will see from the above comments I have reached out to a several appropriate reviewers with only one responding so far.

I will try a few more today, but please do include any suggestions you may have.
You can explore the [reviewer database](https://reviewers.joss.theoj.org/) to make suggestions, bearing in mind that they need to conform to the [JOSS CoI Policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)
