### 2025-08-12 ff3f73e
WIP gee + cdse + tests
