+++ terragon/microsoft_planetary_computer.py
@@ -1,5 +1,4 @@
 import shutil
-from collections import defaultdict
 from typing import Union
 from urllib.parse import urljoin
 
@@ -11,7 +10,7 @@ import xarray as xr
 from joblib import Parallel, delayed
 
 from .base import Base
-from .utils import meters_to_crs_unit
+from .utils import gather_assign_meta, meters_to_crs_unit
 
 
 class PC(Base):
@@ -107,6 +106,8 @@ class PC(Base):
         res = meters_to_crs_unit(self._param("resolution"), shp)
 
         if self._param("create_minicube"):
+            # order items by time because odc.stac.load does not preserve the order (preserve_original_order does not work)
+            items = sorted(items, key=lambda x: x.datetime)
             ds = odc.stac.load(
                 items,
                 bands=self._param("bands"),
@@ -116,25 +117,7 @@ class PC(Base):
                 y=(bounds[1], bounds[3]),
                 **self._param("odc_stac_kwargs", default={}),
             )
-            if (
-                isinstance(self._param("save_metadata"), (list, tuple))
-                and len(self._param("save_metadata")) > 0
-            ):
-                # create a dict with metadata: {"prop1": [value1, value2, ...], ...}
-                meta_dict = defaultdict(list)
-                # loop through items and collect metadata
-                for item in items:
-                    item = item.to_dict()
-                    for key in self._param("save_metadata"):
-                        if key in item:
-                            meta_dict[key].append(item[key])
-                        elif key in item["properties"]:
-                            meta_dict[key].append(item["properties"][key])
-                        else:
-                            raise ValueError(f"Key {key} not found in item properties or features.")
-                # assign metadata to time dimension as coords
-                for key, values in meta_dict.items():
-                    ds = ds.assign_coords({key: ("time", values)})
+            ds = gather_assign_meta(self, items, ds)
             ds = self._prepare_cube(ds)
             return ds
         else:
