+++ tests/copernicus_data_space_ecosystem.py
@@ -70,8 +70,15 @@ class TestCDSE(_TestBase, unittest.TestCase):
         args["end_date"] = "2022-01-04"
         args["resolution"] = 10
         args["filter"] = {"productType": {"eq": "IW_GRDH_1S-COG"}}
+        args["save_metadata"] = ["id"]
 
         ds = self.tg.create(**args)
+
+        # check meta data
+        self.assertTrue("id" in ds.coords)
+        times = ds.time.dt.strftime("%Y%m%d")
+        ids_dates = [id.split("_")[5].split("T")[0] for id in ds.id.values]
+        self.assertTrue(all(times == ids_dates))
 
         # test if the dimensions are correct, with +/- 1 pixel distance
         self.assertTrue(
@@ -278,8 +285,21 @@ class TestCDSE(_TestBase, unittest.TestCase):
         args["end_date"] = "2021-01-16"
         args["resolution"] = 30
         args["filter"] = {"processingLevel": {"eq": "LEVEL2"}}
+        args["save_metadata"] = ["id"]
 
         ds = self.tg.create(**args)
+
+        # check meta data
+        self.assertTrue("id" in ds.coords)
+        times = ds.time.dt.strftime("%Y%m%d")
+        ids_dates = [id.split("_")[3] for id in ds.id.values]
+        self.assertTrue(all(times == ids_dates))
+
+        self.assertTrue(len(ds.time) == 4 and len(ds.x) == 10 and len(ds.y) == 7)
+
+        self.assertTrue(
+            all(times == ids_dates)
+        )
 
         self.assertTrue(len(ds.time) == 4 and len(ds.x) == 10 and len(ds.y) == 7)
 
