+++ terragon/utils.py
@@ -1,4 +1,5 @@
 import warnings
+from collections import defaultdict
 
 import pyproj
 from shapely.geometry import Point
@@ -123,3 +124,31 @@ def meters_to_crs_unit(meters, shp):
     distance_units_y = Point(orig_point).distance(Point(offset_point_in_orig_crs_y))
 
     return distance_units_x, distance_units_y
+
+
+def gather_meta(items: dict, meta_names: list):
+    # create a dict with metadata: {"prop1": [value1, value2, ...], ...}
+    meta_dict = defaultdict(list)
+    # loop through items and collect metadata
+    for item in items:
+        if not isinstance(item, dict):
+            item = item.to_dict()
+        for key in meta_names:
+            if key in item:
+                meta_dict[key].append(item[key])
+            elif key in item["properties"]:
+                meta_dict[key].append(item["properties"][key])
+            else:
+                raise ValueError(f"Key {key} not found in item properties or features.")
+    return meta_dict
+
+
+def gather_assign_meta(obj, items: dict, ds):
+    if (
+        isinstance(obj._param("save_metadata"), (list, tuple))
+        and len(obj._param("save_metadata")) > 0
+    ):
+        meta_dict = gather_meta(items, obj._param("save_metadata"))
+        # assign metadata to time dimension as coords
+        ds = ds.assign_coords({key: ("time", values) for key, values in meta_dict.items()})
+    return ds

[…]
