### 2025-08-13 25451bf
WIP new cdse class for stac
