### 2025-08-13 7297f17
WIP new cdse class for stac
