+++ terragon/copernicus_data_space_ecosystem.py
@@ -1,0 +1,147 @@
+import os
+from urllib.parse import urljoin
+
+import odc.stac
+import pystac_client
+import requests
+
+from .base import Base
+from .utils import gather_assign_meta, meters_to_crs_unit
+
+
+class CDSE(Base):
+    """Class to interact with the Copernicus Data Space Ecosystem. The images are downloaded via the STAC API.
+    Check the official CDSE website for the supported collections: https://documentation.dataspace.copernicus.eu/APIs/newSTACcatalogue.html
+
+    :param credentials: credentials to authenticate, expected format: {'aws_access_key_id': <id>, 'aws_secret_access_key': <key>}.
+    :param base_url: the URL for the STAC catalog, defaults to "https://stac.dataspace.copernicus.eu/v1/"
+    """
+
+    def __init__(
+        self,
+        credentials: dict,
+        base_url: str = "https://stac.dataspace.copernicus.eu/v1/",
+    ):
+        """Initialize class and save the credentials.
+
+        :param credentials: credentials to authenticate, expected format: {'aws_access_key_id': <id>, 'aws_secret_access_key': <key>}.
+        :param base_url: the URL for the STAC catalog, defaults to "https://stac.dataspace.copernicus.eu/v1/"
+        :raises ValueError: when the credentials are in the wrong format
+        """
+        super().__init__()
+        self.base_url = base_url
+        if credentials:
+            if "aws_access_key_id" not in credentials or "aws_secret_access_key" not in credentials:
+                raise ValueError(
+                    "aws_access_key_id or aws_secret_access_key not in credentials, could not initialize."
+                )
+        else:
+            # fallback to credentials saved in the environment/files from aws
+            pass
+
+        # only needed for direct-stac # TODO
+        os.environ["GDAL_HTTP_TCP_KEEPALIVE"] = "YES"
+        os.environ["AWS_S3_ENDPOINT"] = "eodata.dataspace.copernicus.eu"
+        os.environ["AWS_HTTPS"] = "YES"
+        os.environ["AWS_VIRTUAL_HOSTING"] = "FALSE"
+        os.environ["GDAL_HTTP_UNSAFESSL"] = "YES"
+        os.environ["AWS_ACCESS_KEY_ID"] = credentials.get("aws_access_key_id", None)
+        os.environ["AWS_SECRET_ACCESS_KEY"] = credentials.get("aws_secret_access_key", None)
+        self.credentials = credentials
+
+    def retrieve_collections(self, filter_by_name: str = None):
+        """Search the collections provided by Copernicus Data Space Ecosystem.
+
+        :param filter_by_name: name to filter the collections for, defaults to None
+        :raises RuntimeError: if the request to the collections endpoint fails
+        :return: a list of collection names
+        """
+        collections_url = urljoin(self.base_url, "collections")
+        response = requests.get(collections_url)
+        if response.status_code == 200:
+            data = response.json()
+            collections = [collection["id"] for collection in data["collections"]]
+            if filter_by_name:
+                collections = [
+                    collection for collection in collections if filter_by_name in collection.lower()
+                ]
+            return collections
+        else:
+            raise RuntimeError("Failed to retrieve collections")
+
+    def search(
+        self,
+        *args,
+        **kwargs,
+    ):
+        """Search for items in the Copernicus Data Space Ecosystem collections via stac. For a description of the args/kwargs parameters see the Base class function.
+
+        :raises ValueError: when no items are found or parameters are in the wrong format
+        :raises RuntimeError: when the corresponding files for the items are not found
+
+        :return: a list of items
+        """
+        super().search(*args, **kwargs)
+
+        # use pystac_client to search for items
+        catalog = pystac_client.Client.open(self.base_url)
[…]
