### 2025-08-13 7773e7a
add documentation and demo
