+++ docs/source/cdse_s3.rst
@@ -1,0 +1,6 @@
+Copernicus Data Space Ecosystem S3 (CDSE_S3) Class
+============================================
+.. automodule:: terragon.copernicus_data_space_ecosystem_s3
+   :members:
+   :undoc-members:
+   :show-inheritance:
