### 2025-08-19 554c104
Adding metadata save
