### 2025-08-19 e6f3e2f
Small fix with time sorting
