+++ terragon/alaska_satellite_facility.py
@@ -1,7 +1,6 @@
 import logging
 import shutil
 import warnings
-from collections import defaultdict
 from pathlib import Path
 
 import asf_search as asf
@@ -14,7 +13,7 @@ from rasterio.vrt import WarpedVRT
 from shapely.geometry import box
 
 from .base import Base
-from .utils import align_coords, align_resolutions
+from .utils import align_coords, align_resolutions, gather_assign_meta
 
 
 class ASF(Base):
@@ -235,25 +234,9 @@ class ASF(Base):
 
         if self._param("create_minicube"):
             ds = self._create_minicube(items)
-            # Create dict with metadata
-            if (
-                isinstance(self._param("save_metadata"), (list, tuple))
-                and len(self._param("save_metadata")) > 0
-            ):
-                # create a dict with metadata: {"prop1": [value1, value2, ...], ...}
-                meta_dict = defaultdict(list)
-                # loop through items and collect metadata
-                for item in items:
-                    item = item.properties
-                    for key in self._param("save_metadata"):
-                        if key in item:
-                            meta_dict[key].append(item[key])
-                        else:
-                            raise ValueError(f"Key {key} not found in item properties or features.")
-                # assign metadata to time dimension as coords
-                for key, values in meta_dict.items():
-                    ds = ds.assign_coords({key: ("time", values)})
-
+            ds = gather_assign_meta(
+                self._param("save_metadata"), [item.properties for item in items], ds
+            )
             ds = self._prepare_cube(ds)
 
             if self._param("rm_tmp_files"):
