### 2025-08-21 29ce9fa
feature: collect meta data as coordinate  (#28)
