### 2025-08-21 a530a27
feature: collect meta data as coordinate  (#28)
