### 2025-08-21 0d7e19f
fix config (mocking) for docs
