+++ docs/source/asf.rst
@@ -1,5 +1,5 @@
 Alaska Satellite Facility (ASF) Class
-===============================
+=====================================
 .. automodule:: terragon.alaska_satellite_facility
    :members:
    :undoc-members:
