+++ terragon/__init__.py
@@ -1,3 +1,3 @@
 from .init import init  # noqa: F401
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
