### 2025-08-24 #8139 @jatkinson1000
Hi there @AtmaMani, @lwasser, @davemlz, and @DS4Ag!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a toolkit that provides a common interface to access several Earth Observation datasets allowing users to generate flexible data-polygons.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
