### 2025-08-24 #8139 @jatkinson1000
@drnhhl, @usethedata and @AtmaMani I am going to start the review process now.
This will take place in a separate issue where there will be instructions and guidance and you will be automatically assigned.

In general JOSS prefers to have three reviewers on papers, but given the difficulty I have had in finding people I am going to get the review process started now that we have two and perhaps add another at a later date -- If anyone requested above would like to review please post a comment and I will add you to the review.
