### 2025-09-09 drnhhl/terragon#30 @AtmaMani
`pip` install with optional dependencies does not work on zsh due to hyphen
Running `pip install terragon-downloader[pc]` or `pip install terragon-downloader[gee]` returns error on zsh shells, but not on bash or other flavors:

```
zsh: no matches found: terragon-downloader[gee]
```
Since `zsh` is a fairly popular shell variant, I suspect other users would run into it. A simple fix is to quote the package name as below:

```
pip install "terragon-downloader[pc]"
```
which will not make the shell expand any characters.
