### 2025-09-09 drnhhl/terragon#31 @AtmaMani
[DOC] - Lazy loads vs actual downloads is not clear
The library uses Joblib and integrates with xArray which is good. I notice that when I create the xArray DA object, only pointers to the dataset is loaded (lazy loads). The data is loaded only upon actual usage (be it query or rendering). I think this needs to be communicated to the user - particularly, which action gets the data.
