### 2025-09-09 drnhhl/terragon#32 @AtmaMani
[DOC] Licenses of upstream data providers is not clear
The library is essentially a discover and use tool for multiple satellite data vendors. Laborious tasks like account creation, portal specific queries and downloads are all simplified. However, it is not clear if the consumer gets to read / accept the data vendor's licensing terms and conditions.

It is likely as you add more vendors to the mix, each comes with their own usage and limitations. I feel it is good to think about this now and document that.

On a simple level it could be: Ask the user to visit GEE / PC / Alaska portal, create an account and agree to terms and conditions.
On an elaborate level, it could be: Find a mechanism to present those via the library and get the agreement.
