### 2025-09-09 drnhhl/terragon#34 @AtmaMani
[DOC] - Remove the doc for base.py
Since the `Base` class in `base.py` is abstract and not meant to be instantiated or used by the end user, I suggest removing it from the sphinx API builder queue
