### 2025-09-09 drnhhl/terragon#35 @AtmaMani
[DOC] Non deterministic nature of query results
Since the library is at the mercy of upstream APIs by different vendors, some text about this reality is needed. This is important for reproducible research - a command may give different results based on what those vendors deprecate. The user should be advised that such a thing could happen and is beyond the control of this library
