### 2025-09-09 drnhhl/terragon#36 @AtmaMani
[DOC] - Expand documentation on data discovery feature
This library builds on top of vendor REST APIs, STAC catalogs, STAC APIs. This is great and I fully support it. However, I'd like to see a top level page that explains how to look for data - expand about the `collections` property
