### 2025-09-09 drnhhl/terragon#37 @AtmaMani
[FEAT] Provide metadata and summary about collections
I  like the many examples presented in the doc, but in most, the user already has to know what collection to query. For someone in the EO space, this is trivial. But for a newcomer, this is a hurdle. I think this library can help with that hurdle.

I present two ideas:

**Simple**: Return the `collections` as a dictionary. In addition to the collection id, add some metadata about it.
**Elaborate**: Return the `collections` as a rich object. In a jupyter notebook env, this can be a list of cards. In simpler REPL environment, this could adapt using the `__display__` or `__repr__` built ins to display appropriately.

Here's an example of what we did for the ArcGIS API for Python: https://developers.arcgis.com/python/latest/guide/accessing-and-creating-content/#searching-for-content-outside-your-portal
