### 2025-09-09 drnhhl/terragon#38 @AtmaMani
[doc] Contribution doc
In the README.md file on the root, contributors are asked to reach out. However, you already have a proper contribution page:  https://terragon.readthedocs.io/en/stable/source/contributing.html. I suggest updating the readme file to point to that link
