### 2025-09-09 #8857 @AtmaMani
@jatkinson1000 I reviewed this paper and the library. It presents a good value for the EO community. I added a few suggestions. I am linking them below:

* https://github.com/drnhhl/terragon/issues/30
* https://github.com/drnhhl/terragon/issues/31
* https://github.com/drnhhl/terragon/issues/32
* https://github.com/drnhhl/terragon/issues/33
* https://github.com/drnhhl/terragon/issues/34
* https://github.com/drnhhl/terragon/issues/35
* https://github.com/drnhhl/terragon/issues/36
* https://github.com/drnhhl/terragon/issues/37
* https://github.com/drnhhl/terragon/issues/38
