### 2025-09-10 9e4cc71
update readme according to #30 #32 #35 #38
