### 2025-09-10 67ddbb8
update readme according to #30 #32 #35 #38
