### 2025-09-10 drnhhl/terragon#35 @drnhhl
Thank you for this suggestion. This is an important topic, and we agree that we should add a warning. Additionally, we have added a warning about the compatibility between the data providers. The text in the README Limitations states now:
"This library relies on external data providers. Therefore, the reproducibility cannot be guaranteed and depends on the providers. The data offered may also vary across providers. It is important to note that each provider operates independently and utilizes different processing pipelines. This can result in various products that may not be compatible with one another, even if their collections share a similar name on the platforms of the data providers. Additionally, certain data or patches may be available from some providers but not from others. This includes mosaicking, which depends on the chosen collection and is not handled by this library.".

Changes made in 9e4cc7158b10b26cefbe8d1b6ac7af7bedc2a0c4 and will be included in the next release.
