### 2025-09-10 drnhhl/terragon#30 @drnhhl
Good catch! We have updated the README accordingly.

Changes made in 9e4cc7158b10b26cefbe8d1b6ac7af7bedc2a0c4 and will be included in the next release.
