### 2025-09-10 drnhhl/terragon#32 @drnhhl
Yes, the users should comply with the data providers' licenses, and users need to create accounts if the data providers have an authentication mechanism. Where to create the license and which authentication data to provide is also described in the Jupyter notebooks for each data provider. However, we should make this clearer. Hence, we have added text to the README stating:
"Users must provide authentication (if required) for the data providers to Terragon and must comply with their licensing agreements. Instructions on how to create accounts and the necessary information for each data provider can be found in the [Demos](https://github.com/drnhhl/terragon/tree/main/docs/demo_files)."

Changes made in 9e4cc7158b10b26cefbe8d1b6ac7af7bedc2a0c4 and will be included in the next release.
