### 2025-09-11 drnhhl/terragon#31 @drnhhl
Thank you for raising this issue. I am not sure I have fully understood your query. The user interacts with two functions: "search" and "download". The first only filters for the items and sets the parameters, while the second actually downloads the data and returns an xarray Dataset, and the data downloading/loading is done in this function. Do you have a concrete example where you only receive pointers or where the dataset is loaded lazy?
