### 2025-09-15 e381bc5
WIP: gather collections with metadata
