+++ docs/requirements-docs.txt
@@ -4,3 +4,5 @@ nbsphinx
 sphinx-rtd-theme
 myst-parser
 sphinx-autodoc-typehints
+ipywidgets
+ipytree

[…]
