+++ terragon/alaska_satellite_facility.py
@@ -94,11 +94,15 @@ class ASF(Base):
 
         return collections
 
-    def search(self, *args, rm_tmp_files=True, resampling=rasterio.enums.Resampling.nearest, **kwargs):
-        """Search for items in the Alaska Satellite Facility collections. For a description of the args/kwargs parameters see the Base class function.
+    def search(
+        self, *args, rm_tmp_files=True, resampling=rasterio.enums.Resampling.nearest, **kwargs
+    ):
+        """Search for items in the Alaska Satellite Facility collections, return the items and their meta data,
+        and store the parameters in the class in order to access them later in the download function.
 
         :param rm_tmp_files: Remove downloaded temporary files after creating the data cube, defaults to True.
         :param resampling: Resampling method to use when reprojecting images, defaults to rasterio.enums.Resampling.nearest.
+        :param args/kwargs: Parameters which are handled by the parent class, these parameters are the same for all data providers. See the 'Base' class for more information.
         :raises ValueError: If no items are found for the given search parameters.
         :return: A list of ASF products (items).
         """
@@ -145,10 +149,10 @@ class ASF(Base):
         return items
 
     def _download_item(self, item, session, output_dir, bands=None):
-        """Download a complete ASF item via HTTP and extract its relevant TIFF files.
+        """Download a complete ASF item via HTTP and extract its relevant TIFF files and return the data as xarray.Dataset.
 
         If the item has already been downloaded and the expected files exist, the download is skipped.
-        Otherwise, the method downloads the zip file, extracts the specified TIFF files, and updates the item.
+        Otherwise, the method downloads the zip file, extracts the specified TIFF files.
 
         :param item: ASF item containing metadata and the file URL.
         :param session: HTTP session to use for the download.
