### 2025-09-16 e3b8d54
remove test from cdse l8 -> was removed from cdse + formatting
