### 2025-09-16 cdac93d
Searchable collections with metadata (#41)
Improvements for #40, #37, #36
