### 2025-09-16 1ca8057
Searchable collections with metadata (#41)
Improvements for #40, #37, #36
