+++ README.md
@@ -28,7 +28,7 @@ We currently support these data providers:
 - [Planetary Computer (pc)](https://planetarycomputer.microsoft.com/catalog)
 - [Google Earth Engine (gee)](https://developers.google.com/earth-engine/datasets)
 - [Copernicus Data Space Ecosystem (cdse)](https://dataspace.copernicus.eu/explore-data/data-collections) (not all collections supported)
-- [Alaska Satellite Facility (asf)](https://asf.alaska.edu/how-to/data-basics/datasets-available-from-asf-sar-daac/) (not all collections supported)
+- [Alaska Satellite Facility (asf)](https://docs.asf.alaska.edu/datasets/using_ASF_data/) (not all collections supported)
 
 ## Usage
 ### Installation
