### 2025-09-16 drnhhl/terragon#36 @drnhhl
Thank you for raising this point. We have not documented that well. To fix this, we included a new feature, see issue #37, and added a new demo page called "Finding Data / Collections". Additionally, in each demo, we have added a short example on how to use it, and we have added the links to the data explorers of the data provider in the demos.

Changes were made in 1ca80574a49357b9424e9be223aa12fa6dcceb49 and will be included in the next release.
