### 2025-09-16 drnhhl/terragon#37 @drnhhl
Yes, we agree it would be valuable to also get some more information about the metadata of the collections. However, our focus is on the downloading capabilities. Hence, we now return a dictionary with the metadata, which is searchable via its fields with a regex, and it can be filtered by fields. We have also expanded the documentation and have added a new demo page, which shows the functions' capabilities, see issue #36.

Changes were made in 1ca80574a49357b9424e9be223aa12fa6dcceb49 and will be included in the next release.
