### 2025-09-16 drnhhl/terragon#34 @drnhhl
We agree that this structure is not necessarily relevant to the user. We have thought about including it when creating the documentation, too. However, we came to the conclusion that it provides the best overview. Otherwise, we would need to document the parameters in the corresponding data provider class and repeat them every time. Right now, each references the base class. This also makes it much clearer which functions and parameters are shared/different between the providers. Hence, we argue for leaving the structure as it is. We have expanded the description in the documentation to make this distinction clearer. Please let me know if you think there’s anything critical we may have overlooked. Otherwise, I will go ahead and close this issue.
