### 2025-09-17 172b6ac
update readme citing
