### 2025-09-17 #8857 @drnhhl
Dear @usethedata,
We thank you very much for your comments and suggestions. Each of the raised points has been addressed in the corresponding issue. Regarding NASA Earth Data support, we are currently working on a solution. Unfortunately, the implementation will require more time and is unlikely to be completed before publication. For the other issue, we have added a disclaimer in the README and decided against implementing the requested feature, as the development costs outweigh the expected benefits. Finally, we emphasize that the end-user remains responsible for ensuring proper citation and should cross-check the citation even if the citation information is provided through our library. We thank you again for your feedback and hope that our responses provide sufficient clarification regarding the raised points.
