### 2025-09-18 drnhhl/terragon#34 @AtmaMani
@drnhhl sounds good. I see your point as well. We can close this issue.
