### 2025-09-18 drnhhl/terragon#31 @AtmaMani
I was under the assumption that xArray DataArray objects were merely pointers to data still residing remotely. For instance, with PC, the data library being STAC compliant and files likely to be COGs, I thought the DataArray objects simply streamed content from COGs still residing in blob store. Hence that question. I inspected the source code and found in `https://github.com/drnhhl/terragon/blob/dev/terragon/microsoft_planetary_computer.py#L144` and similar for other vendors that the library actually download the files.

Where is the (`download_folder`)[https://github.com/drnhhl/terragon/blob/dev/terragon/microsoft_planetary_computer.py#L120] param configured? Is this exposed to the user? If the files are downloaded in a usable format (like geotiff), are those exposed to the user? I as a user would find both the raw files and the earth data cube useful.
