### 2025-09-18 drnhhl/terragon#34 @AtmaMani
@drnhhl the changes made so far (in this and other issues) look good. Do you plan to merge them to main and update the package on PyPI? I am not sure if that is a pre-requisite to clear for JOSS publication. Either way, I recommend doing it and publishing a `v0.3` as I feel the "exploration" component you added/enhanced warrants a dot version bump.
