### 2025-09-18 #8857 @jatkinson1000
Super, thanks @AtmaMani
To answer that question (and for the benefit of @drnhhl) JOSS requires that the 'reviewed software' i.e. the software seen by the reviewers, including any changes made as a result of the review in order to be approved, is made into a tagged release and has a DOI associated with it (more details to follow).
