### 2025-09-19 drnhhl/terragon#31 @drnhhl
The user can choose whether to download files or create a minicube using the parameter [create_minicube](https://github.com/drnhhl/terragon/blob/172b6ac920746850c607d0c65ebbe3d824211f36/terragon/base.py#L47). Additionally, the user can specify a [download_folder](https://github.com/drnhhl/terragon/blob/172b6ac920746850c607d0c65ebbe3d824211f36/terragon/base.py#L45) for the downloads. When a minicube is created, downloading files is generally prohibited, unless it is necessary for some data providers whose APIs only support file downloads, such as GEE. If a user opts to create a minicube, the files will be deleted after the cube is generated, although users have the option to retain these temporary files if they prefer not to redownload the data when creating the minicube multiple times.

For Planetary Computer, we still aim to create the cube without downloading the files. The odc-stac library offers the [option chunks](https://odc-stac.readthedocs.io/en/latest/_api/odc.stac.load.html#odc.stac.load) for using Dask, which would show the behavior you described, but we download the pixels directly. The line you referenced is downloading files when the user opts not to create a minicube. Therefore, users have both options: they can either download a minicube efficiently, or they can choose to download the files without creating a minicube, in which case the download function will return a list of file names instead of a minicube.
