### 2025-09-29 144df1a
Small editorial typographic corrections for JOSS publication.
