+++ paper/paper.md
@@ -28,18 +28,18 @@ bibliography: paper.bib
 ---
 
 # Summary
-Terragon (Earth(Poly)gon) is a Python package facilitating access to remote sensing and Earth observation data from multiple sources. Its goal is to unify the process of downloading data in a simple and efficient manner. While existing tools focus on specific satellites or data providers, Terragon offers a more flexible solution. The package offers a consistent way to search, filter, and download data from various data sources. It utilizes a polygon format to define the region of interest and creates a spatio-temporal data cube [@mahechaEarthSystem2020] of rasterized data in the xarray Dataset format, as illustrated in Figure \ref{workflow}. Additionally, it ensures the alignment of projections and resolutions, organizing the data according to the selected resolution and coordinate reference system.
+Terragon (Earth(Poly)gon) is a Python package facilitating access to remote sensing and Earth observation data from multiple sources. Its goal is to unify the process of downloading data in a simple and efficient manner. While existing tools focus on specific satellites or data providers, Terragon offers a more flexible solution. The package offers a consistent way to search, filter, and download data from various data sources. It utilizes a polygon format to define the region of interest and creates a spatio-temporal data cube [@mahechaEarthSystem2020] of rasterized data in the Xarray Dataset format, as illustrated in Figure \ref{workflow}. Additionally, it ensures the alignment of projections and resolutions, organizing the data according to the selected resolution and coordinate reference system.
 
 ![Terragon Workflow \label{workflow}](image.pdf){height="300pt"}
 
-Currently, several common data providers are supported, including Google Earth Engine [@GORELICK201718], Planetary Computer [@microsoft_open_source_2022_7261897], Copernicus Data Space Ecosystem [@ecosystemCopernicusData2024] and Alaska Satellite Facility [@ASFHome]. The goal is to further develop and maintain this tool by incorporating additional data providers and implementing processing techniques, such as mosaicking and resampling, based on community needs. The software was leveraged to prepare two large-scale datasets Sen12Landslides [@Sen12Landslides_2025] and CropClimateX [@cropclimatex_dataset]. Sen12Landslides contains 75,000 landslide annotations and has over 12,000 patches from Sentinel 1 and 2. CropClimateX contains 15,500 small data cube spanning 1527 counties in the USA, it spans multiple sensors, weather and extreme events, soil and terrain features. More projects are in preparation.
+Currently, several common data providers are supported, including Google Earth Engine [@GORELICK201718], Planetary Computer [@microsoft_open_source_2022_7261897], Copernicus Data Space Ecosystem [@ecosystemCopernicusData2024] and Alaska Satellite Facility [@ASFHome]. The goal is to further develop and maintain this tool by incorporating additional data providers and implementing processing techniques, such as mosaicing and resampling, based on community needs. The software was leveraged to prepare two large-scale datasets Sen12Landslides [@Sen12Landslides_2025] and CropClimateX [@cropclimatex_dataset]. Sen12Landslides contains 75,000 landslide annotations and has over 12,000 patches from Sentinel 1 and 2. CropClimateX contains 15,500 small data cube spanning 1527 counties in the USA, it spans multiple sensors, weather and extreme events, soil and terrain features. Further projects are currently in preparation.
 
-Overall, Terragon reduces the resources required for the time-consuming process of accessing and downloading data from various APIs, combining it into a consistent, reusable and cost-effienct framework. 
+Overall, Terragon reduces the resources required for the time-consuming process of accessing and downloading data from various APIs, condensing them into a consistent, reusable and cost-efficient framework. 
 
 # Statement of need
[…]
