### 2025-09-29 2064924
full form for acronyms
