+++ paper/paper.md
@@ -39,7 +39,7 @@ Overall, Terragon reduces the resources required for the time-consuming process 
 # Statement of need
 The number of active Earth-observation satellites is increasing [@SatelliteDatabase], and platforms like Sentinel-2 are generating vast amounts of data. This development has significantly impacted remote sensing analysis by providing access to large-scale datasets, both free and commercial. Simultaneously, there is an increasing demand for analysis-ready data to develop data-driven methods, such as machine learning models. However, the process of gathering this data remains time-consuming. Existing platforms are fragmented, often providing only specific satellites, formats, and custom APIs. This leads to repeated work and poor compatibility across platforms, highlighting the need for more convenient access and a unified, straightforward framework.
 
-Although many useful tools have already been developed to help people access and process Earth observation data [@Taconet2024; @Chudley2024; @Haan2023; @montero2024cubo; @hogenson2020hyp3], these tools often support only one satellite, product, or API, or are built for specific programming languages like R. This limits their broader applicability, especially for researchers looking to combine data from multiple sources or work within a Python-based scientific stack, where Python has become the standard for machine learning research. Although there is some functional overlap with cubo [@montero2024cubo], our design diverges substantially. Cubo builds fixed-size data cubes around a centroid point using UTM zones, while Terragon supports arbitrary polygons as input and preserves the source CRS, avoiding pixel misalignment and overlaps. We aim to have a cube that closely mimics the shape of the polygon, making it a far more flexible approach. Cubo's design does not support this paradigm and would require rewriting its core architecture. 
+Although many useful tools have already been developed to help people access and process Earth observation data [@Taconet2024; @Chudley2024; @Haan2023; @montero2024cubo; @hogenson2020hyp3], these tools often support only one satellite, product, or API, or are built for specific programming languages like R. This limits their broader applicability, especially for researchers looking to combine data from multiple sources or work within a Python-based scientific stack, where Python has become the standard for machine learning research. Although there is some functional overlap with cubo [@montero2024cubo], our design diverges substantially. Cubo builds fixed-size data cubes around a centroid point using Universal Transverse Mercator (UTM) zones, while Terragon supports arbitrary polygons as input and preserves the source coordinate reference system, avoiding pixel misalignment and overlaps. We aim to have a cube that closely mimics the shape of the polygon, making it a far more flexible approach. Cubo's design does not support this paradigm and would require rewriting its core architecture. 
 
 Terragon addresses these challenges, allowing users to focus on their applications rather than the technical setup. This simplifies scaling to different platforms and eliminates the need for repeated setups for each data provider. Terragon offers a straightforward and consistent interface for downloading data from various APIs. It produces clean, analysis-ready data cubes that integrate seamlessly into machine learning workflows and other applications.
 
