### 2025-10-14 f6d0990
update upper case references
