+++ paper/paper.bib
@@ -1,19 +1,19 @@
-@article{Taconet2024, doi = {10.21105/joss.07343}, url = {https://doi.org/10.21105/joss.07343}, year = {2024}, publisher = {The Open Journal}, volume = {9}, number = {103}, pages = {7343}, author = {Paul Taconet and Nicolas Moiroux}, title = {modisfast: An R package for fast and efficient access to MODIS, VIIRS and GPM Earth Observation data}, journal = {Journal of Open Source Software} }
+@article{Taconet2024, doi = {10.21105/joss.07343}, url = {https://doi.org/10.21105/joss.07343}, year = {2024}, publisher = {The Open Journal}, volume = {9}, number = {103}, pages = {7343}, author = {Paul Taconet and Nicolas Moiroux}, title = {modisfast: An {R} package for fast and efficient access to MODIS, VIIRS and GPM {Earth Observation} data}, journal = {Journal of Open Source Software} }
 @article{Chudley2024, doi = {10.21105/joss.07149}, url = {https://doi.org/10.21105/joss.07149}, year = {2024}, publisher = {The Open Journal}, volume = {9}, number = {102}, pages = {7149}, author = {Thomas R. Chudley and Ian M. Howat}, title = {pDEMtools: conveniently search, download, and process ArcticDEM and REMA products}, journal = {Journal of Open Source Software} }
-@article{Haan2023, doi = {10.21105/joss.05205}, url = {https://doi.org/10.21105/joss.05205}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {89}, pages = {5205}, author = {Sebastian Haan and Januar Harianto and Nathaniel Butterworth and Thomas Bishop}, title = {Geodata-Harvester: A Python package to jumpstart geospatial data extraction and analysis}, journal = {Journal of Open Source Software} }
+@article{Haan2023, doi = {10.21105/joss.05205}, url = {https://doi.org/10.21105/joss.05205}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {89}, pages = {5205}, author = {Sebastian Haan and Januar Harianto and Nathaniel Butterworth and Thomas Bishop}, title = {{Geodata-Harvester}: A {Python} package to jumpstart geospatial data extraction and analysis}, journal = {Journal of Open Source Software} }
 @article{montero2024cubo,
   doi = {10.48550/ARXIV.2404.13105},
   url = {https://arxiv.org/abs/2404.13105},
   author = {Montero,  David and Aybar,  César and Ji,  Chaonan and Kraemer,  Guido and S\"{o}chting,  Maximilian and Teber,  Khalil and Mahecha,  Miguel D.},
   keywords = {Databases (cs.DB),  Computer Vision and Pattern Recognition (cs.CV),  Machine Learning (cs.LG),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
-  title = {On-Demand Earth System Data Cubes},
+  title = {{On-Demand Earth System Data Cubes}},
   publisher = {arXiv},
   year = {2024},
   copyright = {Creative Commons Attribution 4.0 International}
 }
 @software{hogenson2020hyp3,
   author       = {Hogenson, K. and Kristenson, H. and Kennedy, J. and Johnston, A. and Rine, J. and Logan, T. and Zhu, J. and Williams, F. and Herrmann, J. and Smale, J. and Meyer, F.},
-  title        = {Hybrid Pluggable Processing Pipeline (HyP3): A cloud-native infrastructure for generic processing of SAR data},
[…]
